requirements.txt
patchright==1.52.5
python-dotenv==1.1.1
openai==1.35.0